hmac: move `EagerHash` to the crate root
